Issue#2; Active DID isn't saved on logout/login
Users table now has a FK reference to the dids table. This points to the active DID of the user, which gets updated at the same time that the session variable does.

This reference can be null, as users could possibly have no dids under their account (or are in the process of syncing their dids).